Break out deleteTasks into helper functions
