+++ packages/geoprocessing/scripts/clear/deleteTasks.ts
@@ -1,17 +1,14 @@
 import { DynamoDBClient } from "@aws-sdk/client-dynamodb";
+import { DynamoDBDocument } from "@aws-sdk/lib-dynamodb";
 import {
-  BatchWriteCommand,
-  BatchWriteCommandInput,
-  DynamoDBDocument,
-  ScanCommandInput,
-  paginateScan,
-  DynamoDBDocumentPaginationConfiguration,
-} from "@aws-sdk/lib-dynamodb";
+  TaskKey,
+  batchDeleteTasks,
+} from "../../src/aws/dynamodb/batchDeleteTasks.js";
+import { scanTasks } from "../../src/aws/dynamodb/scanTasks.js";
+import { wait } from "../../src/helpers/wait.js";
 
 const MAX_BATCH_DELETE = 25; // 25 is the maximum number of items that can be deleted in a single batch
-
-type TaskKey = { id: string; service: string };
-let tableName = "";
+const WAIT_TIME = 2000; // 2 seconds between batch deletes, to avoid throttling
 
 /**
  * Clear all results from task table
@@ -22,45 +19,15 @@ export async function deleteTasks(
   region: string,
   serviceName?: string
 ) {
-  tableName = `gp-${projectName}-tasks`;
-
-  // console.log("projectName", projectName);
-  // console.log("region", region);
-  // console.log("serviceName", serviceName);
-
+  const tableName = `gp-${projectName}-tasks`;
   const docClient = DynamoDBDocument.from(
     new DynamoDBClient({
       region: region,
     })
   );
 
-  // Get all task keys.  Scans entire table, may take a while
+  const pager = scanTasks(docClient, tableName, serviceName);
 
-  const paginatorConfig: DynamoDBDocumentPaginationConfiguration = {
-    client: docClient,
-    pageSize: 25,
-  };
-
-  let query: ScanCommandInput = {
-    TableName: tableName,
-    ProjectionExpression: "id, service",
-  };
-  if (serviceName && serviceName !== "all") {
-    query = {
-      TableName: tableName,
-      ProjectionExpression: "id, service",
-      FilterExpression: "service = :pk",
-      ExpressionAttributeValues: {
-        ":pk": serviceName,
-      },
-    };
-  }
-
-  // Pager will return a variable number of items, up to 1MB of data
-  const pager = paginateScan(paginatorConfig, query);
-
-  //
-  let promises: Promise<void>[] = [];
   let taskKeys: TaskKey[] = [];
   let hasItems = false;
   // let batchNum: number = 0;
@@ -68,7 +35,7 @@ export async function deleteTasks(
   for await (const result of pager) {
     if (result && result.Items && Object.keys(result.Items).length > 0) {
       hasItems = true;
-      result.Items.forEach(async (item, batchNum, result) => {
+      result.Items.forEach(async (item) => {
         taskKeys.push({
           id: item.id,
           service: item.service,
@@ -77,7 +44,8 @@ export async function deleteTasks(
         // When batch of tasks is ready, start their delete and continue
         if (taskKeys.length >= MAX_BATCH_DELETE) {
           const taskKeyBatch = taskKeys.splice(0, MAX_BATCH_DELETE); // deletes items from taskKeys array
-          await batchDeleteTasks(docClient, taskKeyBatch, tableName); // wait for delete of whole batch before continuing
+          await batchDeleteTasks(docClient, taskKeyBatch, tableName);
+          await wait(WAIT_TIME);
         }
       });
     }
@@ -92,109 +60,5 @@ export async function deleteTasks(
     console.log(
       `No results found in DynamoDB table ${tableName} ${serviceName ? "for service " + serviceName : ""}`
     );
-  }
-
-  promises.forEach(async (p) => await p);
-}
-
-/**
- * Batch delete array of tasks
- */
-async function batchDeleteTasks(
-  docClient: DynamoDBDocument,
-  taskKeys: TaskKey[],
-  tableName: string
-) {
-  // const batchSize = taskKeys.length;
-  // const lowerBound = batchNum * MAX_BATCH_DELETE + 1;
-  // const upperBound =
-  //   batchNum * MAX_BATCH_DELETE +
-  //   (batchSize < MAX_BATCH_DELETE ? batchSize : MAX_BATCH_DELETE);
-  // console.log(`Deleting items ${lowerBound} - ${upperBound}`);
-
-  const deleteRequestChunk: object[] = taskKeys.map((taskKey) => {
-    return {
-      DeleteRequest: {
-        Key: {
-          id: taskKey.id,
-          service: taskKey.service,
-        },
-      },
-    };
-  });
-
-  const deleteRequest: BatchWriteCommandInput = {
-    RequestItems: {
-      [tableName]: deleteRequestChunk,
-    },
-  };
-
-  await batchDelete(docClient, deleteRequest, 0, 10);
-  await wait(2000);
-}
-
-const wait = (ms: number) => new Promise((res) => setTimeout(res, ms));
-
-async function batchDelete(
-  docClient: DynamoDBDocument,
-  deleteCommandInput: BatchWriteCommandInput,
-  retryCount: number = 0,
-  maxRetries: number = 10
-) {
-  try {
-    // console.log("deleting", JSON.stringify(deleteCommandInput, null, 2));
-    const id =
-      deleteCommandInput!.RequestItems![tableName!][0].DeleteRequest!.Key!.id;
-    const service =
-      deleteCommandInput!.RequestItems![tableName!][0].DeleteRequest!.Key!
-        .service;
-    console.log(
-      `${retryCount > 0 ? "Retry #" + retryCount + " " : ""}Deleting batch starting with ${id} - ${service}`
-    );
-    const deleteCommand = new BatchWriteCommand(deleteCommandInput);
-    const res = await docClient.send(deleteCommand);
-
-    if (res.UnprocessedItems && Object.keys(res.UnprocessedItems).length > 0) {
-      if (retryCount > maxRetries) {
-        throw new Error(
-          `${Object.keys(res.UnprocessedItems).length} items not deleted after ${maxRetries} retries`
-        );
-      }
-
-      const id = res.UnprocessedItems[tableName!][0].DeleteRequest!.Key!.id;
-      const service =
-        res.UnprocessedItems[tableName!][0].DeleteRequest!.Key!.service;
-      // console.log(
-      //   `  ${Object.keys(res.UnprocessedItems[tableName]).length} unprocessed, retry batch in ${2 ** retryCount * 10}ms, starting with ${id} - ${service}`
-      // );
-      await wait(2 ** retryCount * 10); // wait time increases exponentially
-
-      await batchDelete(
-        docClient,
-        { RequestItems: res.UnprocessedItems }, // call again with unprocessed items
-        retryCount + 1,
-        maxRetries
-      );
-    }
-  } catch (e: any) {
-    // console.log(JSON.stringify(e, null, 2));
-    if (
-      e.$metadata &&
-      e.$metadata.httpStatusCode === 400 &&
-      e.$metadata.totalRetryDelay
-    ) {
-      const id =
-        deleteCommandInput!.RequestItems![tableName!][0].DeleteRequest!.Key!.id;
-      const service =
-        deleteCommandInput!.RequestItems![tableName!][0].DeleteRequest!.Key!
-          .service;
-      // console.log(
-      //   ` ThroughputError, retry in ${e.$metadata.totalRetryDelay}ms starting with ${id} - ${service}`
-      // );
-      await wait(e.$metadata.totalRetryDelay);
-      await batchDelete(docClient, deleteCommandInput, 0, maxRetries);
-    } else {
-      throw new Error(e);
-    }
   }
 }
