+++ packages/geoprocessing/src/helpers/wait.ts
@@ -1,0 +1,1 @@
+export const wait = (ms: number) => new Promise((res) => setTimeout(res, ms));
